+++ src/structvisualizer/Controller.java
@@ -8,6 +8,11 @@ import javafx.fxml.Initializable;
 import javafx.scene.control.*;
 import javafx.scene.layout.*;
 import parser.SomeClass;
+import structvisualizer.animatecollections.AnimateStructure;
+import structvisualizer.animatecollections.AnimateStructureFactory;
+import structvisualizer.data.Types;
+import structvisualizer.valuefactories.ListValuesFactory;
+import structvisualizer.window.ErrorWindow;
 
 import java.net.URL;
 import java.util.ResourceBundle;
