+++ src/structvisualizer/Window/InputClassWindow.java
@@ -1,4 +1,4 @@
-package structvisualizer;
+package structvisualizer.window;
 
  /*
  * InputClassWindow   3/22/16, 15:36
@@ -9,24 +9,20 @@ package structvisualizer;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import javafx.collections.ObservableListBase;
 import javafx.geometry.*;
-import javafx.geometry.Insets;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
-import javafx.scene.control.TextArea;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
-import javafx.scene.text.Text;
-import javafx.scene.text.TextFlow;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
 import parser.SomeClass;
+import structvisualizer.Main;
+import structvisualizer.data.Types;
 
-import java.net.URI;
 import java.util.HashMap;
 
 /**
