+++ src/structvisualizer/data/Collections.java
@@ -1,4 +1,4 @@
-package structvisualizer;
+package structvisualizer.data;
 
 /**
  * Enum for supported structures of data
@@ -6,7 +6,7 @@ package structvisualizer;
  * @author Kyrylo Havrylenko
  * @see
  */
-interface Collections {
+public interface Collections {
     final String ARRAY_LIST = "ArrayList";
     final String STACK = "Stack";
     final String HASH_MAP = "HashMap";
