+++ src/structvisualizer/data/Types.java
@@ -1,4 +1,4 @@
-package structvisualizer;
+package structvisualizer.data;
 
 /**
  * Enum inteface for supported types
